aprendendo sobrecarga de metodos

+++ src/cursojava/javacore/Bintroducaometodos/dominio/Teste_Algoritmos.java
@@ -2,7 +2,7 @@ package cursojava.javacore.Bintroducaometodos.dominio;
 
 public class Teste_Algoritmos {
 
-    public double calculator (float num1,int num2) {
+    public int calculator (int   num1,int num2) {
         return num1 + num2;
 
     }
